Move PenaltyDirichletBC from tests and add solid_mechanics test closes #5268
Also made the penalty a required parameter.  Seems to me that if
you're using penalty anything, you'd better know what the penalty
is.
