+++ modules/solid_mechanics/tests/penalty_dirichlet/tests
@@ -1,0 +1,7 @@
+[Tests]
+  [./test]
+    type = 'Exodiff'
+    input = 'penalty_dirichlet.i'
+    exodiff = 'out.e'
+  [../]
+[]
