+++ test/tests/bcs/penalty_dirichlet_bc/function_penalty_dirichlet_bc_test.i
@@ -56,6 +56,7 @@
     variable = u
     function = solution
     boundary = 'top left right bottom'
+    penalty = 1e6
   [../]
 []
