+++ test/tests/kernels/coupled_kernel_grad/coupled_kernel_grad_test.i
@@ -104,6 +104,7 @@
     variable = u
     function = slnu
     boundary = 'left right top bottom'
+    penalty = 1e6
   [../]
   [./bc_v]
     type = FunctionDirichletBC
@@ -117,6 +118,7 @@
     variable = u
     function = slnu
     boundary = 'left right top bottom'
+    penalty = 1e6
   [../]
   [./bc_u_tb]
     type = CoupledKernelGradBC
