+++ test/tests/misc/check_error/coupled_grad_without_declare.i
@@ -107,6 +107,7 @@
     variable = u
     function = slnu
     boundary = 'left right top bottom'
+    penalty = 1e6
   [../]
   [./bc_v]
     type = FunctionDirichletBC
@@ -120,6 +121,7 @@
     variable = u
     function = slnu
     boundary = 'left right top bottom'
+    penalty = 1e6
   [../]
   [./bc_u_tb]
     type = CoupledKernelGradBC
